[chore #161532537] Rename Register to User, and add validation for role
