+++ server/src/schema/migration.js
@@ -32,7 +32,6 @@ CREATE TABLE IF NOT EXISTS users(
     username VARCHAR(100) NOT NULL,
     password VARCHAR(100) NOT NULL
     role VARCHAR(100) NOT NULL
-    
 );
       `);
   },
